Some customizations for Z shell
I started to migrate from using bash to using tmux and zsh. Thus, this
is doing some small customizations to this new (for me) terminal
enviromnent:

* Adding tmux.conf. For now, it only has one line to enable 256 colors
on the terminal because, by default, it was enabling only 8 colors.
Having only 8 colors available in the terminal was quite limiting when
doing customizations.
* Updating vimrc, so that vim colors are OK in tmux, under the assuption
of a dark backround in the terminal.
* Adding zshrc file with some initial customizations of the shell
(custom prompt, enabling direnv, history options, syntax hightlighting,
enable default completions, and keybindings for easier transition from
bash).
* Adding zsh-syntax-highlighting submodule, because that's what provides
zsh hightlighting.

This is also synching existent dotfiles which had small adjustments:

* Synching bashrc, because HISTCONTROL option was added.
* Adding the gemrc from the work laptop, because it has reasonable flags
for the ruby `gem` command.