Alias podman command as docker
This is because some of the source code I work with have hard references
to docker command and don't let easily change the command.

Fedora has podman in its repositories which can be used as a drop in
replacement for docker. So, add some lines to install-container-tools.sh
to create a symlink to alias the podman command as docker. This is done
only if docker CLI is not available.