Move minikube installation to individual script
Rather than installing minikube as part as the environment setup script,
this creates a new helper func `app-install` to install it. The new
individual script is written with the feasibility for upgrading to a
newer version.

This new `app-install` func will be the entry point for installing other
apps/tools and following the same idea of upgradeability support. The
reason for extracting to individual script rather than keeping in the
main setup script is that not all tools are required when configuring a
new machine. Also, not all tools need to be upgraded at the same time.
Thus, likely this on-demand approach is better.

This extraction will be done gradually. Other basic tooling may be
installed in the main setup script, although may still be moved to an
independent script (for upgradeability) being called from the main setup